Adding hash verification; test refactor
Here, I'm adding a hash verification as the last step in a successful
jlap patch operation. The `patch_repo_data` function now returns this
updated hash. I'm guessing it will be useful for updating the
`*.state.json` file.

This commit also refactors the tests a little bit to make it easier to
reason about (trying to hide some of the setup code in their own
functions).