Updates to return the hash object instead of string
This commit also makes changes to the rattler_digest code. I think it's
better to store the hash type there rather than in the
rattler_repodata::fetch::cache module that is private.

I also updated the way that the hashes are generated to take advantage
of the rattler_digest library too.