Lots of refactor from manual testing
This commit inclues a lot of refactoring I did based on manual testing
with real repodata (still needs to be included as unit tests 😬). The
case I was not handling well were empty JLAP responses. This happens
when there is no data to update. I was also saving the wrong values for
the new initialization vector and had to address this too.