refactor: remove SolvableMetadata
The SolvableMetadata was kind of a hack to add certain associations.
This was only used to lookup the original package record associated with
a solvable. However, since the refactor this information is directly
stored in the solvable so we can just extract the information from the
pool directly. This simplifies the code and removes the need for
additional bookkeeping.