feat: display merged candidates (#326)
feat: adds functionality that gives the dependency provider the option
to display the merged candidates for errors in a way it sees fit

A default display provider is given. The conda code overwrites this
provider

Also removes `VersionTrait` as we no longer need this. Removed this both
from the test and conda code