fix: jlap wrong hash (#390)
Fixes an issue where the client-side hash was used to determine the
required server-side patches. This will fail in a lot of cases because
the data is not actually bit-for-bit the same. JLAP actually stores the
server-side hash separately, which is not what this code does as well. I
also refactored the code a bit to remove whitespaces in the repodata
file which saves a significant amount of space.