make builds deterministic (#392)
This was just for testing. 
I think we might want some more control.

The `append_file` function wasn't actually using the prepared header so
the `mtime` was still set from the file system. Changing that to the
`tar::HeaderMode::Deterministic` works. I think we might still want more
control. The Deterministic header from the tar library does something to
the `chmod` bits and also sets the mtime to the same constant that I
copied out because that is the birthdate of Rust and some programs don't
handle a mtime of `0` not well apparently.