add options to disable zstd and bz2 (#420)
We want to try `rattler` and `rattler-build` within the client
infrastructure. This client uses JFrog Artifactory to proxy conda
repositories. The deployed version has a bug where the
`repodata.json.zst` is cached indefinitely. An option to disable support
for bz2 and zstd would allow us to use rattler with this version.