fix wildcard expansion for stored credentials of domains (#442)
While testing the `auth-file` feature for `rattler-build`
(https://github.com/prefix-dev/rattler-build/pull/413) I noticed that
the wildcard extension for domains does not work correctly:

`https://repo.prefix.dev` was matched with `repo.prefix.dev` and
`*.repo.prefix.dev`, but not with `*.prefix.dev`. On closer inspection,
it looks like the function never worked properly: See
https://github.com/mamba-org/rattler/pull/252

This PR fixes the wildcard handling and adds a unit test for it.