feat: refactor `Version.bump()` to accept bumping `major/minor/patch/last` (#452)
See https://github.com/mamba-org/rattler/issues/449 for context.

- [x] add new enum `VersionBumpType`
- [x] refactor `Version.bump()` to `Version.bump(bump_type:
VersionBumpType)`
- [x] add unit tests
- [x] adapt the python binding

This PR is breaking backward compat but since rust does not support
overloading or default args, I could not find a more creative way to
keep backward compat. That being said, it might not be that much of a
problem. If it is, please let me know, and I can always keep `bump()` as
it is and create a new function.

---------

Co-authored-by: Tarun Pratap Singh <101409098+Wackyator@users.noreply.github.com>