fix: consistent clobbering & removal of `__pycache__` (#437)
This attempts to fix two issues:

- remove empty `__pycache__` folders for `noarch` packages (otherwise
the package is still "detected" in the `site-packages` folder).
- do not attempt to link individual files more than once per
installation. This happens when there are "clobbering" files. Currently
a random file will win (due to async execution).

However, there is still two remaining issues with clobbering:

- We should deterministically decide on a "winner". I think it should be
by topologically sorting the packages and take the last (top) package as
a winner.
- When removing packages we also need to take care of the
(un-)clobbering. This is a bit more tricky because we might not even
have the clobbering package in the package cache anymore. However, I
think a workaround could be to still link the files into the environment
but with the name of the source package appended (e.g. `crates2.json`
would become `crates2.json-clobber-from-mypackage`. For unclobbering we
could just rename the file.