chore: try to upgrade most dependencies (#492)
- The Zip Reader now reads 3 ranges instead of two, but I didn't
investigate further why that would happen ... kinda weird!
- I didn't manage to update `axum`. They removed the `axum::Server` and
point to `hyper` instead. However, if I see correctly, the server in
hyper is also more complicated. There is a more simple `axum::serve`
that I used e.g. in `rattler-server` but idk how to deal with the
graceful shutdown etc. that we are using.