fix: allow multiple clobbers per package (#526)
There was a bug where we were writing the new prefix records too early.
That made the clobbering resolution only work for the first clobbered
file.

Before, other files are resolved and renamed properly, but this was not
stored in the `PrefixRecord`. Subsequent operations were thus failing.

This modifies the `PrefixRecord` in memory and only writes the new data
at the end of the `unclobbering` operation.