fix: remove drop-bomb, move empty folder removal to `post_process` (#519)
We observed a race condition when removing empty folders - during the
execution of a transaction it can happen that one package gets unlinked
and cleans up empty folders after itself, while another one is already
installing new files into the folders. This is a race condition due to
the way we parellely link all the files.

In order to fix this issue this PR moves the cleaning out of empty
folders until after we're done unlinking and linking new files into the
prefix and then does a (synchronous) cleaning of all empty folders in
the prefix (that are not explicitly mentioned as empty directories).