fix: correct condition to downweigh track-feature packages (#545)
We use track features to de-prioritize packages.
Sometimes, the solver is forced to select a given package, e.g. when
selecting `numpy-XY-<hash>` it might have a dependency on `pypy`. Numpy
variants themselves are not down-prioritized, though.

If we detect that a package forces only dependencies that have track
features (such as the case for the numpy variant that only works
together with `pypy`) we also "inherit" the down-weighing to `numpy`
(otherwise the solver would be forced to use pypy later on).

We already had the code for this but it was missing a `!` – this PR
corrects that.