fix: multi-prefix replacement in binary files (#570)
The `pkg-config` package encodes (in the binary) a "PATH" variable that
looks something like:

```
/prefix_placeholder/lib:/prefix_placeholder/etc/share\0
```

So far we only replaced the initial prefix placeholder, and skipped past
the rest.

With this change we first get the whole "c-string" (until the `\0`), and
then replace each occurence of the prefix in the c-string. The padding
is then added to make sure that the strings are of the same length.