fix: implement cache for authentication filestorage backend (#573)
We've seen issues in pixi where downloading becomes really slow because
the `~/.rattler/auth.json` file is on a network file system.

We also wanted to give better error messages if we cannot parse the JSON
file.

With this change, we cache the authentication information for the
lifetime of the middleware. I think this should fix the two issues
mentioned above.