feat: make root dir configurable in channel config (#602)
Enables resolution of channels relative to the `pixi.toml`, instead of current working directory.  It also makes the call to current work dir the responsibility of the user and they can catch errors (e.g. the current work dir is not available in wasm32 and maybe some other embedded systems or so).