fix: sha computation of symlinks was failing sometimes (#641)
Sometimes linking was failing because we tried to compute the SHA hash
of a symlink-file. When computing the hash, we follow the symlink. But
since we are symlinking everything at once, and in parallel, the
symlinked file is sometimes not yet there.

In that case, the SHA hash computation was failing. 

With this PR, we change it so that we always store `000...000` for
symlinks.
Conda and mamba 'wait' for the files to be ready and then compute (or
find) the hash. We could also do that, but on the other hand the "sha"
hash of a symlink doesn't make that much sense anyways.