perf: create directories up front (#533)
When linking packages a lot of time is spent creating directories as
this flamegraph highlights:


![image](https://github.com/mamba-org/rattler/assets/4995967/90c34e8c-4d4f-4f37-a3a0-3aec531952e5)

Most of the time is spent checking if a directory (or its parents)
already exists. However, we already know upfront all the directories
that need to be created so why not do that? Well, say no more! That is
exactly what this PR does!

New situation:


![image](https://github.com/mamba-org/rattler/assets/4995967/9e47f348-835d-4da8-a8d8-7c8a9476614b)

Its completely gone. Great success! 👍

In my preliminary dead simple benchmarks, this can easily save 20% of
installation times.