fix: make topological sorting support fully cyclic dependencies (#678)
In case the graph of dependencies only consisted of a directed cycle,
`sort_topologically` returned an empty list. This was an issue when
trying to sort the dependencies of an environment containing Python 3.12
and Pip on MacOS as both depend on each other. 

Even though topological sorting theoretically requires a DAG as its
input, the existing implementation of `sort_topologically` already
handled directed cycles just not the case where all the nodes were part
of such cycles.

The fix is that if we can't find any root nodes, we start cycle
detection from a random (the first in the given list) node.