fix: lenient and strict parsing of equality signs (#738)
This PR introduces strict and lenient parsing of versions in the form of
`pytorch=*=*cuda` and `python ==2.7.8.*.*`.

This changes the behavior in strict mode and makes sure that
`NamelessMatchSpec` and `MatchSpec` behave the same way while parsing.

In strict mode only spaces are allowed to separate the package name,
version, and build string. The form `pytorch=*=*cuda` is no longer
allowed and will result in an error. Instead this should be written as
`pytorch * *cuda`. This removes any ambiguity about what the equality
signs mean. Another form that has been removed in strict mode is the
removal of the `==` sign in some specific cases like `python ==2.7`.
Finally, the behavior where the version string is appended with a `*` if
the version starts with a `=` is removed in strict mode.