fix: Fix GatewayQuery.query to filter records based on provided specs (#756)
Fixes #755.

The `query` method of `GatewayQuery` was not filtering records based on
the provided specs. It would only filter records to be fetched when
doing a recursive query.

This PR fixes that.

Note that I'm learning rust and this might not be the right way to fix
this. Please let me know if I can improve this! I'd like to learn.

---------

Signed-off-by: Jean-Christophe Morin <jcmorin@anaconda.com>