fix: constraints on virtual packages were ignored (#795)
If there was a package that had a constraint on a virtual package, e.g.:

```json
"constrains": ["__cuda >=12.1"],
```

and the virtual dependency was never explicitly required it was silently
ignored. So even though your system would only support `cuda 11.8` the
solver could select a version of a package that required a higher
version.

This PR solves this by adding explicit requirements on each specified
virtual package. This behavior is also added for libsolv.