fix: topological sort when cycles appear in leaf nodes (#879)
There were two issues with our topological sort implementation:

- cycles between two noarch packages or two non-noarch packages were
never "broken". This is fixed by this PR (going to add a test next)
- When the root node (or leaf node) contained a cycle, it was not part
of the result. The code is changed to first detect _any and all_ cycles
in the graph, and then find the root nodes with these cycles in mind.

The changes came up because a user was installing `holoviews` which has
a circular dependency on `panel`. It failed later in the un-clobbering
(because some packages were just missing after the topological sort
because of this faulty "root" detection).