fix:  Release downloader lock before re-attempting fallback download  (#938)
### Description
In https://github.com/conda/rattler/pull/837 locking logic was added to
the download reporter to prevent multiple downloads.

This was clashing with the logic added in
https://github.com/conda/rattler/pull/797 where, if we detect we're
trying to download a zip that has data descriptors, we fallback to doing
a full download and interrupt the streaming decompression download.

The issue comes from us not releasing the lock before attempting the
second download, hitting this line
https://github.com/conda/rattler/blob/8bb883536df52de46fd2c040228bb77d39c92607/crates/rattler_cache/src/package_cache/mod.rs#L439

As a fix, we report the previous download as completed before attempting
the second one! Have verified in my custom setup that this works.