feat: merge pixi-build branch (#950)
* feat: lock file update for source dependencies (#866)

Co-authored-by: Bas Zalmstra <zalmstra.bas@gmail.com>
Co-authored-by: Bas Zalmstra <bas@prefix.dev>

* feat: add input hash to conda package

* feat: add input hash/glob to lock file (#868)

* fix: mark records as changed if hash Option value is different (#874)

This adds the case that where the hash changed, which happened for my
source build, that a re-install is triggered. I think this is the
correct behavior, not sure if it will trigger *too* many re-installs but
did not feel that would be the case.

* fix: python bindings issues (#872)

* fix: normalize channel in lock file (#871)

* fix: dead-lock issue

* feat: use sha in cache lock file, needed for source builds. (#900)

* fix: package hash revalidation (#904)

* refactor!: lock-file API (#942)

This PR significantly refactors the lock-file API to make the
distinction between source and binary conda packages. It also changes
the way we handle borrowing data from the lock-file, preferring
borrowing over owned objects.

The Python bindings have also been completely refactored to be more
pythonic. Instead of having a single `LockedPackage` there is no a class
hierarchy of `LockedPackage` -> `CondaLockedPackage`,
`PypiLockedPackage` -> `CondaLockedSourcePackage` and
`CondaLockedBinaryPackage`.

---------

Co-authored-by: Tim de Jager <tdejager89@gmail.com>
Co-authored-by: Tim de Jager <tim@prefix.dev>
Co-authored-by: nichmor <nmorkotilo@gmail.com>