Move files to .trash folder if they are in use
On Windows attempting to delete a file which is in use results in a
permission denied error. However it is possible to rename the file, so
move the files to a folder target_prefix\.trash instead for later
cleanup.

The files are renamed to UUID.rattler_trash to avoid having to worry
about handling filename collisions.