fix: Quote values in bash activation with shlex
On Windows, activating an environment with Git Bash fails, because paths
are being set with double-quote strings: export A="C:\Windows\Path"
By using shlex::try_quote, these variables can be set correctly.

Because setting the PATH used the primitive for setting an environment
variable, that had to change to explicitly use double-quotes, so that
the included ${PATH} is evaluated instead of handled literally.

Extended the tests for ShellScript<Bash> to include '\', an unexpanded
${} variable, all the modes of setting the PATH, and running a script
with a character that needs escaping.