rattler_lock: Add API to read a lock file with a work directory
To support reading relative paths using pep508_rs, we need to
provide a work directory as a base for the relative paths.

Add an API to pass in such an extra directory.

This complicates parsing as serde does not handle that. So aprse
the relevant data into a String using serde and then post-process
it to the real data values.

Add a test-case for reading such a dependency.